Handle zapping 轻量化检测接入 (#107)
* Handle zapping 轻量化检测接入实现

* Update handle-unittest.cc

- 优化 Handle 未逃逸即直接使用单测，让 death 断言的粒度更细
- 补一个根 handle block 不为空的单测
- 顺手将NR_BLOCKS宏修改为函数

+++ src/handles/handle-scope-implementer.h
@@ -15,7 +15,7 @@ class ObjectVisitor;
 class HandleScopeImplementer {
  public:
   // 一个block可以装载的Address数量
-  static constexpr size_t kHandleBlockSize = 512;
+  static constexpr int kHandleBlockSize = 512;
 
   explicit HandleScopeImplementer(Isolate* isolate);
   ~HandleScopeImplementer();
